do a POST creating new record if passed an unsaved AR object
if the object passed to the best_in_place helper responds true for
`new_record?`, the object's unsaved attribs will be passed to the form
builder and the HTTP method will be changed to POST, thereby creating a
new record with the given attribs.

Once the object is saved, the URL is modified to include the new `id`
from the response, and the method is changed to PUT. This way a record
can be created and then updated without reloading the page.

includes a simplification to the `requestData` method to leverage
`jQuery.param()` instead of a slow/cumbersome string builder.